requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
gunicorn==20.0.4
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.6.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1.1
mkl
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.1
object-detection==0.1
olefile==0.46
opencv-python==4.2.0.34
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.4
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
scikit-image==0.17.2
scipy==1.5.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.6.3
torch==1.1.0
torchvision==0.3.0
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.1.0
